refactor(VSkeletonLoader): convert original functionality
