--- packages/vuetify/src/components/VSkeletonLoader/VSkeletonBone.sass
@@ -1,64 +1,0 @@
-// Imports
-@use '../../styles/settings'
-@use '../../styles/tools'
-@use './variables' as *
-
-.v-skeleton-bone
-  background: $skeleton-text-background
-  overflow: hidden
-  position: relative
-
-  @include tools.rounded($skeleton-heading-border-radius)
-
-  &::after
-    @include tools.absolute(true)
-
-    animation: $skeleton-loading-animation
-    background: $skeleton-background
-    transform: $skeleton-loading-transform
-    z-index: 1
-
-.v-skeleton-bone__actions
-  height: $skeleton-button-height
-  width: 100%
-
-.v-skeleton-bone__avatar
-  height: $skeleton-avatar-size
-  width: $skeleton-avatar-size
-
-  @include tools.rounded($skeleton-avatar-border-radius)
-
-.v-skeleton-bone__button
-  height: $skeleton-button-height
-  width: $skeleton-button-width
-
-  + .v-skeleton-bone__button
-    margin-left: 8px
-
-.v-skeleton-bone__heading
-  height: $skeleton-heading-height
-  width: 45%
-
-  + .v-skeleton-bone__subheading
-    margin-top: 4px
-
-.v-skeleton-bone__subheading
-  height: $skeleton-heading-height * 0.8
-  width: 35%
-
-.v-skeleton-bone__image
-  height: $skeleton-image-height
-
-.v-skeleton-bone__sheet
-  background: white
-
-.v-skeleton-bone__text
-  flex: 1 0 auto
-  height: $skeleton-text-height
-
-  + &
-    margin-bottom: 6px
-
-@keyframes loading
-  100%
-    transform: translateX(100%)
