+++ packages/vuetify/src/components/VSkeletonLoader/VSkeletonLoader.tsx
@@ -2,79 +2,92 @@
 // Styles
 import './VSkeletonLoader.sass'
 
-// Components
-import { VSkeletonBone } from './VSkeletonBone'
-
 // Composables
 
 import { makeElevationProps, useElevation } from '@/composables/elevation'
-import { makeThemeProps, provideTheme } from '@/composables/theme'
+import { makeThemeProps, useTheme } from '@/composables/theme'
 
 // Utilities
-import { genericComponent, useRender, wrapInArray } from '@/util'
+import { genericComponent } from '@/util'
 import type { PropType } from 'vue'
-import { computed } from 'vue'
+import { computed, h } from 'vue'
 
-const rootTypes = ['card']
-
-export const types = {
-  actions: ['button@2'],
-  article: ['heading', 'paragraph'],
+export const rootTypes = {
+  actions: 'button@2',
+  article: 'heading, paragraph',
   avatar: 'avatar',
   button: 'button',
-  card: ['image', 'card-heading'],
-  'card-avatar': ['image', 'list-item-avatar'],
+  card: 'image, card-heading',
+  'card-avatar': 'image, list-item-avatar',
   'card-heading': 'heading',
   chip: 'chip',
-  'date-picker': ['list-item', 'card-heading', 'divider', 'date-picker-options', 'date-picker-days', 'actions'],
-  'date-picker-options': ['text', 'avatar@2'],
-  'date-picker-days': ['avatar@28'],
+  'date-picker': 'list-item, card-heading, divider, date-picker-options, date-picker-days, actions',
+  'date-picker-options': 'text, avatar@2',
+  'date-picker-days': 'avatar@28',
   heading: 'heading',
   image: 'image',
   'list-item': 'text',
-  'list-item-avatar': ['avatar', 'text'],
-  'list-item-two-line': ['sentences'],
-  'list-item-avatar-two-line': ['avatar', 'sentences'],
-  'list-item-three-line': ['paragraph'],
-  'list-item-avatar-three-line': ['avatar', 'paragraph'],
-  paragraph: ['text@3'],
-  sentences: ['text@2'],
-  table: ['table-heading', 'table-thead', 'table-tbody', 'table-tfoot'],
-  'table-heading': ['heading', 'text'],
-  'table-thead': ['heading@6'],
-  'table-tbody': ['table-row-divider@6'],
-  'table-row-divider': ['table-row', 'divider'],
-  'table-row': ['table-cell@6'],
+  'list-item-avatar': 'avatar, text',
+  'list-item-two-line': 'sentences',
+  'list-item-avatar-two-line': 'avatar, sentences',
+  'list-item-three-line': 'paragraph',
+  'list-item-avatar-three-line': 'avatar, paragraph',
+  paragraph: 'text@3',
+  sentences: 'text@2',
+  table: 'table-heading, table-thead, table-tbody, table-tfoot',
+  'table-heading': 'heading, text',
+  'table-thead': 'heading@6',
+  'table-tbody': 'table-row-divider@6',
+  'table-row-divider': 'table-row, divider',
+  'table-row': 'table-cell@6',
   'table-cell': 'text',
-  'table-tfoot': ['text@2', 'avatar@2'],
+  'table-tfoot': 'text@2, avatar@2',
   text: 'text',
 } as const
 
-function processElement (element: string | { type: string, children?: any[] }): any[] {
-  const [item, repeat] = (typeof element === 'string' ? element : element.type).split('@')
-  const itemCount = repeat ? parseInt(repeat, 10) : 1
-  const flatItems = Array(itemCount).fill(flattenTypes(item)).flat()
-  const children = typeof element === 'object' && element.children ? flattenTypes(element.children) : []
-
-  return [...flatItems, ...children]
+function genBone (text: string, children: string[] = []) {
+  return h('div', {
+    class: [
+      'v-skeleton-loader__bone',
+      `v-skeleton-loader__${text}`,
+    ],
+  }, children)
 }
 
-function flattenTypes (key: any): any[] {
-  if (typeof key === 'string') {
-    const value = types[key]
+function genBones (bone: string) {
+  // e.g. 'text@3'
+  const [type, length] = bone.split('@') as [string, number]
+  const generator = () => genStructure(type)
 
-    if (key === value) {
-      return [{ type: key }]
-    }
+  // Generate a length array based upon
+  // value after @ in the bone string
+  return Array.from({ length }).map(generator)
+}
 
-    return wrapInArray(value).flatMap(processElement)
-  }
+function genStructure (type: any) {
+  let children = []
+  const bone = rootTypes[type] || ''
 
-  if (Array.isArray(key)) {
-    return key.flatMap(processElement)
-  }
+  // End of recursion, do nothing
+  /* eslint-disable-next-line no-empty, brace-style */
+  if (type === bone) {}
+  // Array of values - e.g. 'heading, paragraph, text@2'
+  else if (type.includes(',')) return mapBones(type)
+  // Array of values - e.g. 'paragraph@4'
+  else if (type.includes('@')) return genBones(type)
+  // Array of values - e.g. 'card@2'
+  else if (bone.includes(',')) children = mapBones(bone)
+  // Array of values - e.g. 'list-item@2'
+  else if (bone.includes('@')) children = genBones(bone)
+  // Single value - e.g. 'card-heading'
+  else if (bone) children.push(genStructure(bone))
 
-  throw new Error('Invalid argument type for flattenTypes function')
+  return [genBone(type, children)]
+}
+
+function mapBones (bones: string) {
+  // Remove spaces and return array of structures
+  return bones.replace(/\s/g, '').split(',').map(genStructure)
 }
 
 export const VSkeletonLoader = genericComponent()({
@@ -85,7 +98,6 @@ export const VSkeletonLoader = genericComponent()({
     loading: Boolean,
     type: {
       type: [String, Array] as PropType<string | string[]>,
-      validator: (val: string) => rootTypes.includes(val),
       default: 'card',
     },
 
@@ -94,35 +106,15 @@ export const VSkeletonLoader = genericComponent()({
   },
 
   setup (props, { slots }) {
-    const { themeClasses } = provideTheme(props)
     const { elevationClasses } = useElevation(props)
+    const { themeClasses } = useTheme()
 
-    useRender(() => {
-      return (
-        <div
-          class={[
-            'v-skeleton-loader',
-            themeClasses.value,
-            elevationClasses.value,
-          ]}
-        >
-          { props.type === 'card' && (
-            <VSkeletonBone
-              class="v-skeleton-loader-card"
-              key="card"
-              type="sheet"
-            >
-              <VSkeletonBone type="image" />
-
-              <VSkeletonBone type="heading" />
-            </VSkeletonBone>
-          )}
-
-          { slots.default?.() }
-        </div>
-      )
+    const items = computed(() => {
+      return genStructure(props.type as any)
     })
 
-    return {}
+    return () => h('div', {
+      class: ['v-skeleton-loader', themeClasses.value, elevationClasses.value],
+    }, [items.value, slots.default?.()])
   },
 })
