+++ packages/vuetify/src/components/VSkeletonLoader/_variables.scss
@@ -29,7 +29,7 @@ $skeleton-loader-date-picker-options-avatar-width: 40px !default;
 $skeleton-loader-date-picker-options-padding: 16px !default;
 $skeleton-loader-divider-border-radius: 1px !default;
 $skeleton-loader-divider-height: 2px !default;
-$skeleton-loader-heading-border-radius: settings.$border-radius-root !default;
+$skeleton-loader-heading-border-radius: 12px !default;
 $skeleton-loader-heading-height: 24px !default;
 $skeleton-loader-image-height: 200px !default;
 $skeleton-loader-item-avatar-height: 40px !default;
