+++ packages/vuetify/src/components/VSkeletonLoader/index.ts
@@ -1,2 +1,1 @@
 export { VSkeletonLoader } from './VSkeletonLoader'
-export { VSkeletonBone } from './VSkeletonBone'
